feat: strict typing, input validation and complete unit test coverage for services
