+++ NEXT_STEPS.md
@@ -1,0 +1,189 @@
+# Guía de mejora post-prueba: API de Cajero (Nest.js)
+
+Este documento te guiará paso a paso para completar y profesionalizar tu API del cajero ahora que puedes tomarte el tiempo necesario. El objetivo es llevarla al nivel que se esperaría en un entorno profesional real.
+
+---
+
+## 🔄 Rama base de trabajo
+
+- Siempre parte desde `development`.
+- Crea ramas `feature/`, `fix/` o `test/` según el objetivo del commit.
+
+Ejemplos:
+
+```bash
+# Antes de cada nueva mejora:
+git checkout development
+git pull origin development
+
+# Crear rama para refactorizar tipado y validaciones:
+git checkout -b feature/strict-types
+
+# Al terminar, commit + push:
+git add .
+git commit -m "refactor: strict typing and validation in services and controllers"
+git push origin feature/strict-types
+```
+
+---
+
+## 📊 Objetivo general
+
+1. Tipado y validaciones estrictas.
+2. Completar tests unitarios (servicio y controlador).
+3. Tests end-to-end.
+4. Autenticación realista con usuarios.
+5. Refactor profesional y modular.
+
+En esta fase, el objetivo es llevar al máximo nivel las funcionalidades ya implementadas:
+
+- Crear cuenta
+- Crear tarjeta
+- Validar tarjeta
+- Cambiar PIN
+
+---
+
+## 📆 Orden recomendado
+
+### ✅ 1. Tipado y validaciones estrictas
+
+- Rama: `feature/strict-types`
+- Revisa todos los DTOs, services y controllers relacionados con cuentas y tarjetas.
+- Usa `class-validator` y `class-transformer` si trabajas con clases.
+- Refuerza validaciones lógicas (por ejemplo: `withdrawal_limit` entre 500 y 6000).
+- Asegura que todos los `service` devuelvan tipos consistentes (`DataResponse<T>`).
+
+### 📅 Commit sugerido
+
+```
+refactor: strict typing and validation for accounts and cards
+```
+
+---
+
+### 🎓 2. Completar tests unitarios de servicios
+
+- Rama: `test/unit-services`
+- Asegúrate de cubrir **todos los casos posibles**:
+    - Casos positivos
+    - Errores esperados (ej: "tarjeta ya activa")
+    - Lógica especial (ej: PIN cifrado, límites)
+- Usa `jest.spyOn` para mocks, y `bcrypt` en modo test.
+
+### 📅 Commit sugerido
+
+```
+test: complete service test coverage for cards and accounts
+```
+
+---
+
+### 🌐 3. Tests unitarios de controladores
+
+- Rama: `test/unit-controllers`
+- Usa `@nestjs/testing` con `createTestingModule`
+- Mockea los servicios para aislar controladores
+- Asegura status codes y formatos de respuesta `{ status, message, data }`
+
+### 📅 Commit sugerido
+
+```
+test: add unit tests for card and account controllers
+```
+
+---
+
+### 🌎 4. Tests end-to-end
+
+- Rama: `test/e2e`
+- Usa `supertest` con `app.getHttpServer()`
+- Cubre:
+    - Creación de cuenta
+    - Creación y activación de tarjeta
+    - Cambio de PIN
+
+### 📅 Commit sugerido
+
+```
+test: e2e for account and card creation, activation, and pin change
+```
+
+---
+
+### 🔐 5. Autenticación segura y modelo de usuario
+
+- Rama: `feature/auth-users`
+- Crea modelo `User` con roles (`admin`, `customer`, etc.)
+- Relaciona `User` con `Account` (por ejemplo, una cuenta pertenece a un usuario)
+- Protege endpoints con JWT (`@nestjs/jwt`, guards y estrategias personalizadas)
+- Valida login con credenciales seguras (por ejemplo, email + password o tarjeta + PIN si decides mantener el estilo del cajero)
+- Protege todas las operaciones sensibles (retiro, cambio de PIN, creación de tarjetas)
+
+### 📅 Commit sugerido
+
+```
+feat: secure authentication with users and account linkage
+```
+
+---
+
+## 📚 Extras profesionales
+
+- Usa `dotenv` + Zod para validar `.env` (✔ hecho)
+- Añade `esbuild` o `tsup` si quieres build rápido
+- Cubre errores 100% con `GlobalExceptionsFilter`
+- Refactoriza errores con una clase `AppError`
+- Documenta la API con Swagger (`@nestjs/swagger`)
+
+---
+
+## 👁️ Consejos finales
+
+- Haz **commits pequeños y descriptivos**.
+- No trabajes nunca en `main`, solo mergea cuando esté revisado.
+- Si una rama se vuelve grande, parte el trabajo en subramas.
+- Agrega tests **antes** de cada refactor (garantiza estabilidad).
+- Usa `prettier` y `eslint` con formato estricto antes de cada push.
+
+---
+
+¡Cuando termines esta guía, tu proyecto estará listo para ser mostrado como ejemplo profesional real! ✅
+
+Estoy desarrollando una API de cajero automático en Nest.js con arquitectura modular y validaciones estrictas. He creado una guía detallada (NEXT_STEPS.md) que indica el orden profesional para llevar la API al máximo nivel. Trabajo siempre en ramas partiendo de development, y ahora estoy en la fase feature/strict-types.
+
+Funcionalidades ya implementadas:
+
+Crear cuenta
+
+Crear tarjeta
+
+Activar tarjeta
+
+Cambiar PIN
+
+Progreso actual:
+
+Servicios y controladores refactorizados con tipado estricto.
+
+DTOs validados con class-validator y documentados con @ApiProperty de Swagger.
+
+Respuestas unificadas con los tipos BaseResponse y DataResponse<T> definidos en shared/types.
+
+Uso correcto de @PrimaryGeneratedColumn para UUID, sin necesidad de uuid() manual.
+
+Validadores personalizados movidos a validators/ dentro del módulo correspondiente.
+
+Swagger funcional, respuestas bien tipadas y coherentes.
+
+Ahora quiero continuar con los tests unitarios, empezando por CardService. Necesito:
+
+Estructura profesional de test (card.service.spec.ts).
+
+Cobertura de casos positivos y errores esperados (por ejemplo: tarjeta ya activada, cuenta inexistente, PIN duplicado, etc.).
+
+Buenas prácticas para mocks de repositorios y servicios (jest.mock, jest.spyOn, etc.).
+
+Tipado fuerte también en los tests.
+
+A partir de aquí, guíame paso a paso. Solo quiero cubrir los tests de servicio por ahora, no avanzar a controladores ni e2e aún.
