+++ package.json
@@ -25,6 +25,7 @@
         "@nestjs/core": "^11.0.1",
         "@nestjs/mapped-types": "*",
         "@nestjs/platform-express": "^11.0.1",
+        "@nestjs/swagger": "^11.1.1",
         "@nestjs/typeorm": "^11.0.0",
         "bcrypt": "^5.1.1",
         "class-transformer": "^0.5.1",
@@ -36,7 +37,6 @@
         "reflect-metadata": "^0.2.2",
         "rxjs": "^7.8.1",
         "typeorm": "^0.3.22",
-        "uuid": "^11.1.0",
         "zod": "^3.24.2"
     },
     "devDependencies": {
