+++ src/modules/account/account.service.ts
@@ -3,7 +3,6 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Account } from './entities/account.entity';
 import { Repository } from 'typeorm';
 import { IBAN, CountryCode } from 'ibankit';
-import { v4 as uuid } from 'uuid';
 
 @Injectable()
 export class AccountService {
@@ -13,10 +12,8 @@ export class AccountService {
     ) {}
 
     public async create(): Promise<Account> {
-        const account = this.accountRepo.create({
-            id: uuid(),
+        const account: Account = this.accountRepo.create({
             iban: IBAN.random(CountryCode.ES).toString(),
-            balance: '0',
         });
 
         return this.accountRepo.save(account);
