+++ src/modules/account/dto/account-id-param.dto.ts
@@ -1,6 +1,8 @@
 import { IsUUID } from 'class-validator';
+import { ApiProperty } from '@nestjs/swagger';
 
 export class AccountIdParamDto {
+    @ApiProperty({ format: 'uuid' })
     @IsUUID('4', { message: 'El ID de cuenta debe ser un UUID válido' })
-    accountId: string;
+    readonly accountId: string;
 }
