+++ src/modules/card/card.controller.ts
@@ -3,16 +3,20 @@ import { CardService } from './card.service';
 import { CardIdParamDto } from './dto/card-id-param.dto';
 import { ActivateCardDto } from './dto/activate-card.dto';
 import { ChangeCardPinDto } from './dto/change-card-pin.dto';
+import { ApiTags, ApiOperation } from '@nestjs/swagger';
+import { BaseResponse } from '@/shared/types/base-response.type';
 
+@ApiTags('Cards')
 @Controller('cards/:cardId')
 export class CardController {
     constructor(private readonly cardService: CardService) {}
 
     @Put('activate')
+    @ApiOperation({ summary: 'Activar una tarjeta con un PIN' })
     async activate(
         @Param() { cardId }: CardIdParamDto,
         @Body() activateCardDto: ActivateCardDto,
-    ) {
+    ): Promise<BaseResponse> {
         await this.cardService.activate(activateCardDto, cardId);
 
         return {
@@ -22,10 +26,11 @@ export class CardController {
     }
 
     @Patch('pin')
+    @ApiOperation({ summary: 'Cambiar el PIN de una tarjeta activa' })
     async changeCardPin(
         @Param() { cardId }: CardIdParamDto,
         @Body() changeCardPinDto: ChangeCardPinDto,
-    ) {
+    ): Promise<BaseResponse> {
         await this.cardService.changeCardPin(changeCardPinDto, cardId);
 
         return {
