+++ src/modules/card/dto/card-id-param.dto.ts
@@ -1,6 +1,8 @@
 import { IsUUID } from 'class-validator';
+import { ApiProperty } from '@nestjs/swagger';
 
 export class CardIdParamDto {
+    @ApiProperty({ format: 'uuid' })
     @IsUUID('4', { message: 'El ID de tarjeta debe ser un UUID válido' })
-    cardId: string;
+    readonly cardId: string;
 }
