+++ src/modules/card/dto/change-card-pin.dto.ts
@@ -1,40 +1,24 @@
-import {
-    Matches,
-    Validate,
-    ValidatorConstraint,
-    ValidatorConstraintInterface,
-    ValidationArguments,
-} from 'class-validator';
-
-@ValidatorConstraint({ name: 'NewPinsMatch', async: false })
-class NewPinsMatchConstraint implements ValidatorConstraintInterface {
-    validate(repeatedNewPin: string, args: ValidationArguments) {
-        const { newPin } = args.object as {
-            newPin: string;
-            repeatedNewPin: string;
-        };
-        return repeatedNewPin === newPin;
-    }
-
-    defaultMessage(): string {
-        return 'Los nuevos PIN introducidos no coinciden.';
-    }
-}
+import { Matches, Validate } from 'class-validator';
+import { ApiProperty } from '@nestjs/swagger';
+import { NewPinsMatchConstraint } from '../validators/new-pins-match.validator';
 
 export class ChangeCardPinDto {
+    @ApiProperty({ example: '1111', description: 'PIN actual' })
     @Matches(/^\d{4}$/, {
         message: 'El PIN actual debe contener exactamente 4 dígitos numéricos.',
     })
-    currentPin!: string;
+    readonly currentPin: string;
 
+    @ApiProperty({ example: '1234', description: 'Nuevo PIN' })
     @Matches(/^\d{4}$/, {
         message: 'El nuevo PIN debe contener exactamente 4 dígitos numéricos.',
     })
-    newPin!: string;
+    readonly newPin: string;
 
+    @ApiProperty({ example: '1234', description: 'Repetir nuevo PIN' })
     @Matches(/^\d{4}$/, {
         message: 'El nuevo PIN debe contener exactamente 4 dígitos numéricos.',
     })
     @Validate(NewPinsMatchConstraint)
-    repeatedNewPin!: string;
+    readonly repeatedNewPin: string;
 }
