+++ src/modules/card/services/hashing-service.ts
@@ -3,13 +3,13 @@ import * as bcrypt from 'bcrypt';
 
 @Injectable()
 export class HashingService {
-    private readonly saltRounds = 10;
+    private readonly saltRounds: number = 10;
 
-    async hash(value: string): Promise<string> {
+    public async hash(value: string): Promise<string> {
         return bcrypt.hash(value, this.saltRounds);
     }
 
-    async compare(plain: string, hash: string): Promise<boolean> {
+    public async compare(plain: string, hash: string): Promise<boolean> {
         return bcrypt.compare(plain, hash);
     }
 }
